requirements.txt
pandas==2.2.0
matplotlib==3.8.2
python-telegram-bot[webhooks]==20.8
python-dotenv==1.0.1